- Fixed the issue that was causing crashes when you tried to place one of them. They had 2 problems: first the template_structure_special whose parent was itself, as Phillip pointed out, but second also incorrect actor names. I also made the Carthaginian wall refer to the right parent template.
- Added a "dropsite" aura to mills, farmsteads and civil centers, and
modified the gathering code so you can only gather objects that have
this aura on them from one of these buildings. You can also add a
traits.supply.forageable tag to an entity to make it ignore the
requirement for dropsites and gather it wherever it is. I did this for
hunting, so you can hunt without being near a farmstead.

This was SVN commit r3265.