Initial commit of GUI changes. You're not liable to see much difference (in fact, there's probably less functionality than there used to be).
This is mostly to commit the new icon sheets so Jason can commit changes
directly if he prefers, and because I'm about to get back onto fixing
the command lists (not necessarily ones that actually do things when you
click them, but at least get them updating properly on the screen for
Matei), and want to commit a semi-usable update before I potentially
break it all again.

This was SVN commit r3571.