* Fixed bug in command lists that was causing the first item in a production list to not be displayed.
* Got selection list (shows a list of portraits of the currently
selected units, with health bars) to work, including horizontally
expanding tab (using a helmet tab temporarily to get an idea of the icon
location).
* Todo: Dynamically adjust tab position when horizontal list (tabs need
more blending room); bleed portraits into additional rows above when the
number of entities in the selection exceeds the length of a list.

This was SVN commit r3674.