#A well eye-candied map for testers and staff members to marvel at the new self-shadowing.
Task things around, guys, and check out how cool the shadows look. You
can even task the ships around. What's cool is that the bireme and
trireme use the "double-click" charge attack, just like the foot units.
This would be a very intuitive way to use the "ramming" ability of
ships.

This was SVN commit r3705.