# fix in sound resource management .. as mentioned in http://www.wildfiregames.com/forum/index.php?showtopic=10719&hl= thanks to matei and philip!
bonus: refs #164
tried to alleviate that by moving the OS error bit to the very end,
which usually means it's beyond the end of the little window unless
you're looking for it. does that help?
i'll venture that the OS error *might* be useful in rare cases and it
wouldn't be bad to have in there.

This was SVN commit r4791.