Added binaries and some example data for COLLADA conversion.
Added FCollada 3.02 library files (but Collada*.dll is not built by
default since it doesn't work in VS2003 (because the FCollada*.dll files
use the VC80 CRT)).
Changed Premake to use optimize-speed instead of optimize, since the
latter does /Ox (in MSVC) which is kind of unofficially deprecated.

This was SVN commit r4935.