disable delay-loading of version.dll.
Of the delay-loaded DLLs, version.dll is the only one that is used in
proximity to WoW64 FS redirection, which apparently has a dangerous
interaction with delay-loading. This should prevent #484 (after
update-workspaces + a rebuild), but we can't be sure due to lack of an
x64 Vista or Win7 machine for testing. Can someone please confirm that
it works without requiring the -quickstart workaround?
refs #484

This was SVN commit r7453.