Fixed Greek temple prop point error.
Also: Re-added the B and C variants, by working around the root node
problem with smoke. Added 'smoke' prop nodes to the A, B, &C base meshes
so that the prop meshes for these variants can continue to use the
'root' node.

Fixed some Player 1 color issues in a few maps. More to come.

This was SVN commit r11293.