Disable distance-sorting for most transparent objects
Distance-sorting of the alpha-blend pass of objects like trees and grass
is quite expensive (partly the cost of sorting, partly the cost of being
unable to batch efficiently), and the effect on visual quality is
typically very hard to see (the alpha-test pass does most of the work
for
objects that aren't fully translucent). That's not a good tradeoff, so
turn it off.

Leave distance sorting enabled for e.g. waterfalls since it might make a
more noticeable difference for them.

This was SVN commit r13753.