Set _NO_DEBUG_HEAP=1 in VS2010 projects
The debug heap has a high performance cost, even in release mode,
when running inside a debugger, which is surprising to people trying
to measure performance. It doesn't seem very useful at finding bugs
either.

Add a NoDebugHeap flag to Premake, to set that environment variable
when running in the VS2010 debugger.

This was SVN commit r13901.