Nerf/rebalance siege towers.
- Armor changed from 2h/20p/10c to 5h/20p/2c
- Attack changed from 25p to 10p/5c
- Bonus added: 2x vs structures
- Ranged reduced from 65 to 55
- Elevation bonus of 10 added
- PreferredClasses is now 'Structure'
- Ungarrisoned arrow count decreased to 0 from 2
- Build time increased to 60 from 50
- BuffHeal is now 0 (decreased from 1)
- Health reduced from 800 to 500
- Walking speed reduced to 5 from 6

The idea is for siege towers to be useful for siege, but also for
cleaning out large numbers of units. Previously they were very good at
the latter and useless for the former. They should still be very
powerful units, but not uncounterable and massing siege towers
(previously unbeatable if enough were acquired) should no longer be
feasible.

This was SVN commit r14135.