Buff catapults slightly.
- Add 'Siege' to the RestrictedClasses of towers, wall towers, and
fortresses. A fortress by itself should not be able to shoot down a
catapult, it should require some defense. This also nerfs building a ton
of towers and fortresses, which was previously very difficult to
penetrate and led to long, boring wars of attrition.
- Increase armor from 2h/10p/5c to 2h/30p/10c (matches siege rams).
- Increase attack from 30h/0p/30c to 0h/20p/35c. Would like to add techs
to increase damage further.
- Reduce splash damage to 0h/15p/30c from 40h/0p/40c.
- Increase maximum range from 68 to 75. May increase this more or add
techs.
Motivation is that siege catapults are relatively weak late-game, as
they have great difficulting penetrating fortified positions, leading to
both sides turtling. Additionally, this makes them roughly as strong as
siege rams, having less damage output than rams but having the ability
to attack from a distance. This balances out the civs without rams and
gives the ones with both rams and catapults an additional choice to
make.

This was SVN commit r14174.