Delete help.txt.
I wanted to update it first for some function names I changed.
Then I noticed that nearly all the functions described there don't exist
anymore or have changed. There are much better descriptions at the
places where the functions are declared/implemented.
We should either do it right (and force everyone to update that file for
every change) or don't to it at all. At this stage of development it
doesn't make sense and when we start documenting all the functions at a
central place, there are better locations than this file.

This was SVN commit r14461.