Fixes warnings printed by the new SpiderMonkey.
Sets SelectableInGameSetup to true in all civ json files that didn't
have it set.
Having this value set to undefined (not set) was interpreted as true,
which is quite arbitrary.
With the new SpiderMonkey a warning will be printed if the value is
undefined.
This avoids ambiguity but makes the setting mandatory.

Refs #2372

This was SVN commit r14802.