Changes to English tooltips for consistency
Reported by an Italian translator, Fabio, at Transifex:
https://www.transifex.com/projects/p/0ad/translate/#it/public/13718374
https://www.transifex.com/projects/p/0ad/translate/#it/public/24494405

The periods were added in the main menu (only 2 tooltips without them)
and removed from the options menu (only 1 tooltip with it). In the
future, we might want to be consistent with tooltips through all the
GUI.

This was SVN commit r15052.