Sets default compiler for OS X build to clang, using hack that should be replaced if we upgrade premake.
Fixes linking error in ICU build from build-osx-libs.sh (it was setting
the C++ compiler to clang, which is the C compiler, so it wouldn't
automatically link to the c++ lib). Now build-osx-libs.sh and the game
should use clang and whichever c++ lib is the default

This was SVN commit r15725.