Improve performance when looking for new targets.
Tested in combat demo huge where nearly 50% of the replay time is spent
looking for new targets, and GetPreference is called a lot as used to
sort possible targets.
Caching the preferred classes before the second loop, the replay time of
combat demo huge is reduced by about 15%. Replacing the "for each" loops
by "for of" loops gains an additionnal 7%.

This was SVN commit r16057.