Disable online reporting of profiler data.
Nobody ever looks at the data, and it takes up a lot of server disk
space. The feature should probably be restored in the future once we've
fixed all the obvious performance problems, since a wider range of data
lets us search for system-dependent issues (e.g. some GL drivers are
particularly slow with particular features), but it's not worthwhile
now.

This was SVN commit r16213.