Display all auras in the tooltips (even if there are some with an identical name). Patch by s0600204, fixes #3655.
Add an aura tooltip to the construction panel.
Remove duplicate tooltip code for the training panel.
Eliminate for-each and use the aura ID in GetTemplateDataHelper, since
only that is a unique identifier.
Change the name from "Hero Aura" to "Garrisoned Capture Aura" for that
aura that every hero has.

This was SVN commit r17644.