Improve the resource distribution on the map "Ambush" and maps using bluffs in general. Based on patch by _kali, refs #3455.
Bluffs:
Cover half instead of most of the area of bluffs with forests (to allow
placing CCs / forts there).
Proportionally reduce the likelyhood of placing metal/stone on bluffs.
Compensate the absence of forests on the savanna biome with more
straggler trees.

Ambush:
Place forests also on the mainland (so that players are not immediately
forced to expand to bluffs).
Reduce the amount of berries and animals a bit following 91e8248273.

This was SVN commit r17945.