Session / unit-actions cleanup.
Eliminate deprecated for-each.
Merge if-conditions.
Add linebreaks to objects, so that each property is declared on a
separate line, refs #252.
For consistency, use returns in negative instead of positive cases.
Make three functions global, one of which had been defined thrice.
Remove comments stating the same as the code.

This was SVN commit r18224.