Savegame, session and gamesetup cleanup.
Don't save the player assignments to savegames as
the GUIDs are outdated and the GUI data only applies to the player that
saved the game.
Remove unneeded argument from getPlayerData.
Remove invalid (ready should be status) and unneeded playerAssignment
data when selecting a map in singleplayer mode.
Show filenames of saved games in the debug log.

This was SVN commit r18355.