Trireme and bireme ship balancing.
The mauryan empire is not known to have the strongest fleet, so reduce
their ships hitpoints to the average for that model.
Redue bireme metal cost to make it an actual alternative to the trireme,
considering population cost and hitpoints.
Remove a hardoded linenumber in a comment.

This was SVN commit r18402.