Lobby / gamesetup cleanup.
Translate the map "random" in the lobby.
Fix a warning in getMapDisplayName reported by echotangoecho when
hosting a lobby game and selecting the random map.
Use the right translation of that "random" map in b84b9a374a (since it's
not the maptype).
Notice mapInfoName still shouldn't be colored.

This was SVN commit r18428.