Merge duplicate big house templates. Patch by Sandarac, reviewed by mimo, fixes #4100.
While at it, remove the revert of the
civbonus_hellenic_architecture.json in the hellenic house templates
(1091 Health * 1,1 = 1200 Health),
since the templates should be independent from the number of the json
file (thus restore the missing 10% bonus for hellenic houses).
Construct big houses in 50 (respectively 55) seconds instead of 60
seconds for the same reason (and using a number that is more round than
53 without increasing the buildtime).

This was SVN commit r18517.