Ram garrison balancing.
Garrison up to 10 men in rams and 12 in persian ones, in order to
have a greater chance of attacking enemies with many melee champs
defending,
since the new balancing in alpha 21 requires siege engines to destroy
buildings, see d38adc7cd0 and 27c46a0a68.

Remove undocumented garrisoned healrate of rams introduced in
d1ed264c9b.
That commit introduced it in many places, which were removed in later
commits.

Leave the positive garrisoned healrate of ships and special buildings
for now,
assuming that there are supplies on ships (but not in rams).

Add garrisoned healrate tooltips in the following commit.

This was SVN commit r18528.