Ingame summary screen and allied resource / population tooltips for active players. Patch by Imarok, also reviewed by fatherbushido and discussed with mimo, fixes #3841.
Shows (only) the stats of the current player in the summary screen while
playing.
Also shows those of mutual allies too if the shared vision tech is
researched.
Display the resources and population counts of allies in the tooltips of
the top panel.
Use a sprintf for the pop / limit ratio and remove hardcoded resource
strings in the XML, refs #3934.

This was SVN commit r18533.