Deal ranged damage even if the attacker dies. Patch by LeanderH. Fixes #3610.
Create a Damage system component from the Damage helper object and parts
of
the Attack component. This fixes the issue by making ranged damage
independent
of the attacking entity.

While there fix the issue of damaging all nearby entities in case the
actual
target was not hit, instead of only a single one.

This was SVN commit r18625.