Major GUI stlye and sprite cleanup.
Remove many unused and orphaned entries from common/,
including the entire Wheat theme from more than a decade ago and
most parts of the Stone theme.

Besides being unused, nearly all of the styles were ugly and don't fit
to the existing theme.
No entry should be reintroduced without being actually used somewhere
and if so, use a new subdirectory with the name of the theme.

This was SVN commit r18633.