Make wooden tower upgradable to the defense tower. Reviewed by fatherbushido, siole and Hannibal Barca, fixes #4125.
Let towers cost 50 wood and reduce the stone cost by 50.
The upgrade costs so much that it's cheaper to build a stone tower than
building and upgrading a wooden tower, but still cheaper than building a
wooden and stone tower.
Increase the minimum distance between wooden tower to 60, so that the
upgrade won't circumstance the minDistance of the stone tower.
Add missing newline to the entity limit tooltip of the upgrade component
(refs #2706) that was never visible before.

This was SVN commit r18754.