Random map script cleanup.
Annotate and move the envrionment defaults to the file that contains the
methods to change them.
Remove redundant watercolor entry and add default water tint to a map
that didn't specify it.
Add a comment about the default waterheight pitfall (4391498cf4).
Add missing quotes and spaces, remove unneeded comments.

This was SVN commit r18862.