Remove a hunk with commented out code and a hunk of broken duplicate code from survival of the fittest, refs #3102.
The removed code is redundant with the AlliedVictoryCheck of the
EndGameManager which is called for all victory conditions.
The code is broken because it suggests last man standing mode while
actually not implementing it.
As of #4143, players can chose between allied victory or last man
standing mode. They should be able to chose on this map too.

This was SVN commit r18906.