Don't spawn units for defeated players and players who won on survival of the fittest. Refs #3102.
The previous code errored as the trigger point became gaia, consequently
trying to access the undefined playerCivicCenter of gaia.
Remove two useless loops by sending the command for all entities instead
of one command for each entity.
Rename attackerEntities to attackerTemplates and attackerCount to
attackersPerTemplate.

This was SVN commit r18913.