Don't use black for gaia on survival of the fittest, refs #3102.
Some GUI elements use white until the first diplomacy change.
The choice black is confusing, problematic as there is a dark gray
player color and
problematic as black words in the chat have no contrast to the black
background.
Changing playercolors in running games is a questionable feature, thus
don't attempt to inform the GUI via a GUIInterface notification now.

This was SVN commit r18914.