Random map fortress generation code cleanup.
Create target objects in a single statement where possible, to highlight
the absence of self-reference.
Separate arrays of hardcoded templates from the logic.
Use loops to remove duplicate statements.
Add scopes to the new code to show and enforce that some variables are
indeed local.

This was SVN commit r18952.