Add text linting to Arcanist.
This introduces a .arclint file used by Arcanist to run a handful of
checks on patches submitted to Phabricator. To keep it portable and
avoid annoying contributors, most of the linting will happen through
Jenkins. This commit uses built-in text linting that checks trailing
whitespace and end of files.

Reviewed By: leper
Differential Revision: https://code.wildfiregames.com/D6
This was SVN commit r19502.