Prohibit cheats if cheats are disabled,
i.e. developer cheats to control enemy units, revealing the map and
promoting units too.

Open the developer overlay only if cheats are enabled (which is always
the case in singleplayer mode and only the case in multiplayer mode if
explicitly enabled).
(This doesn't make it harder for developers to "debug", since they
previously had to remember to disable the rated game setting too. Also
every bug had to be reproduced with the replay either way)

Reverts 3cbbf31173 which reverted 6e0e607b20 (since as of c288278229,
the autostart gamesetup enables cheats too).

Differential Revision: https://code.wildfiregames.com/D455
Fixes #3551
Reviewed By: echotangoecho
Agreed with Itms and Imarok in the last staff meeting.

This was SVN commit r19558.