Women Attack balancing fixes.
Allow (not only spartan but) all women to attack (so that it's more
obvious that they can be used to destroy siege engines if no military
units are available).
(Proposed by scythetwirler in irc on 2017-03-13.)

Remove the celtic woman (3.5x!) attack bonus and their (+50%) health
bonus (they had since the introduction of the britons in 52af3369ba).
(Because woman health is most relevant in the beginning of the game and
at that stage celts already have the population bonus advantage.)

Set Spartan women to Passive instead of Standground, so that they flee.
Convert Spartan woman 2.5x siege bonus to a general +50% Attack bonus
and nerf their 80% HP bonus to +40% (i.e. 40HP instead of 25HP).
Keep Spartan woman ability to patrol and build towers for now.

Differential Revision: https://code.wildfiregames.com/D722
Patch By: borg-
Reviewed By: Grugnas
This was SVN commit r19885.