Purge all History strings from entity templates.
This is so when we mark them for translation (as we will do soon) our
wonderful translators are not inundated with a couple hundred new
strings, many of which will need to be changed
in some way.

Patch by: fatherbushido
Change agreed upon by: elexis
Prep work for: D297
Refs: #3212

This was SVN commit r19960.