Modern style cleanup.
Unify three textPanel style copies in a new ModernTextPanel theme,
following the Credits page in bcf2c75513 and the Splash Screen in
5ce5ef9286 and the Manual in 4b06ebd52d.

Remove the unused textcolor_selected and the (because of a source/gui
bug) defacto unused font property.
Rudimentary scrolling support for the Splash Screen.
Remove now empty files for the time being.

This was SVN commit r19992.